Create Tires Market.md
As of 2024, the global tire market is estimated to be worth US$ 161,790 million, with projections indicating a rise to US$ 204,870 million by 2031, growing at a CAGR of 3.5% from 2025 to 2031. The tire production value in China was US$ 38,945.8 million in 2022, expected to reach US$ 60,246.2 million by 2029, reflecting a 5.92% growth rate from 2023 to 2029. Meanwhile, the North American tire production value stood at US$ 29,728.3 million in 2022, anticipated to rise to US$ 37,754.3 million by 2029 at a growth rate of 4.28%.